Update to Composer 2
This is the second attempt at this. The first one was a failure because
the images used to launch the tests and run the tools used in the
pipeline did not have all the PHP extensions required by the production
code. With Composer 2 this is an issue because there is now a runtime
check to make sure the environment matches the expectations set in the
composer.json.

Note that merging this change will force contributions older than
git #tuleap/stable/d28d5cf3eec949da93ce738b44f6ec08944d9a25 to be
rebased.

Part of request #17910: Compatibility with Composer 2